# `ultralytics 8.3.58` Use `uint8` type for TensorRT Profile