@UltralyticsAssistant 2024-12-20
👋 Hello @Laughing-q, thank you for submitting an `ultralytics/ultralytics` 🚀 PR! To ensure a seamless integration of your work, please review the following checklist:

- ✅ **Define a Purpose**: Clearly explain the purpose of your fix or feature in your PR description, and reference any relevant issues. Ensure your commit messages are clear, concise, and adhere to the project's conventions.
- ✅ **Synchronize with Source**: Confirm your PR is synchronized with the main branch. If it's behind, update it by clicking the 'Update branch' button or by pulling in the latest changes locally.
- ✅ **Ensure CI Checks Pass**: Verify that all Continuous Integration (CI) checks are passing. If any checks fail, investigate and address the issues.
- ✅ **Update Documentation**: Update any relevant documentation for new or modified features.
- ✅ **Add Tests**: If applicable, include or update tests to cover your changes, and confirm that the test suite is passing.
- ✅ **Sign the CLA**: If this is your first contribution, ensure you have signed the Contributor License Agreement (CLA). Add "I have read the CLA Document and I sign the CLA" in a new comment to confirm.
- ✅ **Minimize Changes**: Keep changes to the minimum necessary for the fix or feature update. _"Clarity is the key to success."_ 🧠

An Ultralytics engineer will review your PR soon. If you have any questions, feel free to ask here. Thank you for contributing to Ultralytics! 🚀