@Laughing-q 2024-12-20
@glenn-jocher Updating `uint8` so our profile function compatible with classify task, because of this preprocess:
https://github.com/ultralytics/ultralytics/blob/0504d3a8e004abf068d5c7b7bb3670602d6aa1ef/ultralytics/models/yolo/classify/predict.py#L46
I saw a comment there `must be FP32`, do you remember why it has to be FP32 type?
I tested a detection model profile with/without this PR, and the output speed are identical.
```python
from ultralytics.utils.benchmarks import ProfileModels

results = ProfileModels(["yolo11n.pt"], trt=True, half=True).profile()
```
main:
```bash
| yolo11n            | 640 | - | 18.8±0.2 ms | 0.7±0.0 ms | 2.6 | 6.5 |
```
PR:
```bash
| yolo11n            | 640 | - | 18.8±0.2 ms | 0.7±0.0 ms | 2.6 | 6.5 |
```