@UltralyticsAssistant 2025-01-05
🎉 The PR is merged! Huge thanks to @Laughing-q for spearheading this improvement and to @glenn-jocher for their collaboration. 🚀 Your update to the YOLO benchmark utility is a thoughtful step forward, aligning efficiency with real-world practicality for classification tasks. The future of accurate and impactful ML tools shines that much brighter because of your efforts. 🌟

As Marcus Aurelius said, *"What we do now echoes in eternity."* Your contributions today will undoubtedly ripple through the community, empowering others to achieve even greater things. 💡 Thank you for making a difference!