@Petros626 2025-01-23
@Laughing-q does this only belongs for classification tasks? What about the `must be FP32`?