@Laughing-q 2025-01-24
@Petros626 no worries we tested other tasks profiling with uint8 and also work.